[driver/cc1depscanProtocol] Pass a `DepscanSharing` parameter to the `ScanDaemon` functions, instead of just the `CASOpts`
This provides visibility to other `DepscanSharing` state for `ScanDaemon`,
and also has the benefit of avoiding the need to generate the CAS arguments twice.
